Move sqlite3.h to headers
"sqlite3.h" is needed by header files Binder.h and Extractor.h. Since it
was not in the same directory before, a system file was used instead
which might be a different version than sqlite3.c.

Additionaly: Appply POCO_UNBUNDLED switch consequently in all related files.
