+++ Data/SQLite/src/SQLiteStatementImpl.cpp
@@ -23,7 +23,7 @@
 #if defined(POCO_UNBUNDLED)
 #include <sqlite3.h>
 #else
-#include "sqlite3.h"
+#include "Poco/Data/SQLite/sqlite3.h"
 #endif
